@@ -1,33 +1,119 @@
 [build-system]
-requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+requires = [ "hatchling" ]
 
 [project]
 name = "virtuals-sdk"
-version = "0.1.2"
-authors = [
-  { name = "Your Name", email = "your.email@example.com" },
-]
+version = "0.1.3"
 description = "Official Python SDK for Virtuals API/GAME"
 readme = "README.md"
-requires-python = ">=3.8"
+keywords = [ "python" ]
+authors = [ { name = "Chethan UK", email = "chethanuk@outlook.com" } ]
+requires-python = ">=3.9,<4.0"
 classifiers = [
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
-    "Development Status :: 3 - Alpha",
-    "Intended Audience :: Developers",
-    "Topic :: Software Development :: Libraries :: Python Modules",
+  "Intended Audience :: Developers",
+  "Programming Language :: Python",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 dependencies = [
-    "typing-extensions>=4.0.0",
-    "requests>=2.26.0",
+  "requests>=2.32.3",
+  "typing-extensions>=4.12.2",
 ]
 
-[project.urls]
-"Homepage" = "https://github.com/Virtual-Protocol/virtuals-python"
-"Bug Tracker" = "https://github.com/Virtual-Protocol/virtuals-python/issues"
+urls."Bug Tracker" = "https://github.com/Virtual-Protocol/virtuals-python/issues"
+urls.Documentation = "https://Virtual-Protocol.github.io/virtuals-python/"
+urls.Homepage = "https://Virtual-Protocol.github.io/virtuals-python/"
+urls.Repository = "https://github.com/Virtual-Protocol/virtuals-python"
+
+[dependency-groups]
+dev = [
+  "deptry>=0.21.2",
+  "mkdocs>=1.6.1",
+  "mkdocs-include-markdown-plugin>=7.1.2",
+  "mkdocs-material>=9.5.49",
+  "mkdocstrings[python]>=0.27",
+  "mypy>=1.14.1",
+  "pre-commit>=2.20",
+  "pytest>=8.3.4",
+  "pytest-cov>=6",
+  "responses>=0.25.3",
+  "ruff>=0.8.4",
+  "tox-uv>=1.17",
+]
+
+[tool.hatch.build.targets.wheel]
+packages = [ "src/virtuals_sdk" ]
+
+[tool.ruff]
+target-version = "py39"
+line-length = 120
+fix = true
+
+format.preview = true
+lint.select = [
+  # flake8-builtins
+  "A",
+  # flake8-bugbear
+  "B",
+  # flake8-comprehensions
+  "C4",
+  # mccabe
+  "C90",
+  # pycodestyle
+  "E",
+  # pyflakes
+  "F",
+  # isort
+  "I",
+  # pygrep-hooks
+  "PGH",
+  # ruff
+  "RUF",
+  # flake8-bandit
+  "S",
+  # flake8-simplify
+  "SIM",
+  # flake8-debugger
+  "T10",
+  # pyupgrade
+  "UP",
+  "W",
+  # flake8-2020
+  "YTT",
+  # tryceratops
+  # "TRY",
+]
+lint.ignore = [
+  # LineTooLong
+  "E501",
+  # DoNotAssignLambda
+  "E731",
+]
+lint.per-file-ignores."tests/*" = [ "S101" ]
+
+[tool.pytest.ini_options]
+testpaths = [ "tests" ]
+
+[tool.coverage.report]
+skip_empty = true
+
+[tool.coverage.run]
+branch = true
+source = [ "src/virtuals_sdk" ]
+
+[tool.mypy]
+files = [ "src/virtuals_sdk" ]
+disallow_untyped_defs = true
+disallow_any_unimported = true
+no_implicit_optional = true
+check_untyped_defs = true
+warn_return_any = true
+warn_unused_ignores = true
+show_error_codes = true
