@@ -1,11 +1,12 @@
 import os
+
 from virtuals_sdk import game
 
 agent = game.Agent(
     api_key=os.environ.get("VIRTUALS_API_KEY"),
     goal="search for best songs",
     description="Test Description",
-    world_info="Test World Info"
+    world_info="Test World Info",
 )
 
 # applicable only for platform twitter
@@ -17,13 +18,7 @@ agent.add_custom_function(
     game.Function(
         fn_name="custom_search_internet",
         fn_description="search the internet for the best songs",
-        args=[
-            game.FunctionArgument(
-                name="query",
-                type="string",
-                description="The query to search for"
-            )
-        ],
+        args=[game.FunctionArgument(name="query", type="string", description="The query to search for")],
         config=game.FunctionConfig(
             method="get",
             url="https://google.com",
@@ -31,7 +26,7 @@ agent.add_custom_function(
             platform="twitter",
             success_feedback="I found the best songs",
             error_feedback="I couldn't find the best songs",
-        )
+        ),
     )
 )
 
